fix(test): move Test_Pools_List in pool_create_destroy_test so it is executed before deleting the test pool
